% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prorata.R
\name{prorata}
\alias{prorata}
\title{Prorate values}
\usage{
prorata(y, x, type = c("squared", "absolute"), f = NULL, ...)
}
\arguments{
\item{y}{A double matrix (observations x categories).}

\item{x}{A double array (observations x categories x clusters).}

\item{type}{Optimization type of `"squared"` (default) or `absolute"`.
Ignored if `f` is specified.}

\item{f}{passed to [constrOptim()].}

\item{...}{passed to [constrOptim()].}
}
\value{
A double vector of weights with `prorata` class
}
\description{
Prorate values
}
